- implements full house category - move some test from application to dedicated category test

+++ src/main/java/org/codingdojo/yatzy1/categoryManagement/category/FullHouseCategory.java
@@ -10,19 +10,19 @@ public class FullHouseCategory implements Category {
     @Override
     public int calculate(Dice[] diceFromRoll) {
         Map<Dice, Integer> occurenceDiceMap = CategoryUtil.getOccurenceDiceMap(diceFromRoll);
-        int scoreThreeOfAKind = (int) occurenceDiceMap.keySet().stream()
+        int scoreThreeOfAKind = occurenceDiceMap.keySet().stream()
             .filter(dice->occurenceDiceMap.get(dice) == NumberCategory.THREE_OF_A_KIND)
             .mapToInt(dice->dice.getValue() * NumberCategory.THREE_OF_A_KIND)
             .findFirst()
             .orElse(0);
 
-        int scoreOfPair = (int) occurenceDiceMap.keySet().stream()
+        int scoreOfPair = occurenceDiceMap.keySet().stream()
             .filter(dice->occurenceDiceMap.get(dice) == NumberCategory.PAIR)
             .mapToInt(dice->dice.getValue() * NumberCategory.PAIR)
             .findFirst()
             .orElse(0);
 
-        int scoreOfYatze =  (int) occurenceDiceMap.keySet().stream()
+        int scoreOfYatze =  occurenceDiceMap.keySet().stream()
             .filter(dice->occurenceDiceMap.get(dice) == NumberCategory.YATZY)
             .mapToInt(dice->dice.getValue() * NumberCategory.YATZY)
             .findFirst()
